Update the PIC16F87x examples

+++ targets/Microchip/PIC16F88x/examples/886_signal/main.v
@@ -1,13 +1,13 @@
 // Project name:
-// Author: 
-// Date: 
+// Author:
+// Date:
 // PIC16F886 fosc = 10Mhz
 
 import time
 import port
 
-signal := [1, 2, 4, 8, 16, 32, 64, 128,	//rising
-		   64, 32, 16, 8, 4, 2, 1]	//falling
+signal := [u8(1), 2,  4,  8, 16, 32, 64, 128,	//rising
+		          64, 32, 16, 8, 4,   2, 1]	//falling
 
 port.setup(port.b, port.all_outputs)
